requirements.txt
blinker==1.6.2
click==8.1.3
Flask==2.3.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
psycopg2==2.9.6
python-dotenv==1.0.0
redis==4.5.5
uWSGI==2.0.21
Werkzeug==2.3.6